Limit array length to uint32
Even in wasm64 mode, JS arrays are still bounded by uint32 lengths and indicies.

This works around a current bug in wasm64 mode of Embind.

+++ system/include/emscripten/val.h
@@ -644,11 +644,11 @@ struct BindingType<val> {
 
 template <typename T, typename... Policies>
 std::vector<T> vecFromJSArray(const val& v, Policies... policies) {
-  const size_t l = v["length"].as<size_t>();
+  const uint32_t l = v["length"].as<uint32_t>();
 
   std::vector<T> rv;
   rv.reserve(l);
-  for (size_t i = 0; i < l; ++i) {
+  for (uint32_t i = 0; i < l; ++i) {
     rv.push_back(v[i].as<T>(std::forward<Policies>(policies)...));
   }
